Start a common integration test framework
Allows the database & config initialisation code for
integration testingto be reused by other repositories.
