+++ digitalearthau/testing-default.conf
@@ -1,0 +1,3 @@
+[datacube]
+db_hostname: localhost
+db_database: dea_integration
